fix: don't create devcontainer if it exists

+++ helpers/devcontainer.ts
@@ -46,13 +46,16 @@ export const writeDevcontainer = async (
   framework: TemplateFramework,
   frontend: boolean,
 ) => {
-  console.log("Adding .devcontainer");
+  const devcontainerDir = path.join(root, ".devcontainer");
+  if (fs.existsSync(devcontainerDir)) {
+    console.log("Template already has a .devcontainer. Using it.");
+    return;
+  }
   const devcontainerContent = renderDevcontainerContent(
     templatesDir,
     framework,
     frontend,
   );
-  const devcontainerDir = path.join(root, ".devcontainer");
   fs.mkdirSync(devcontainerDir);
   await fs.promises.writeFile(
     path.join(devcontainerDir, "devcontainer.json"),
